reorganize and rename helper functions
