+++ src/neuroconv/tools/nwb_helpers/__init__.py
@@ -1,0 +1,8 @@
+from ._metadata_and_file_helpers import (
+    add_device_from_metadata,
+    get_module,
+    get_default_nwbfile_metadata,
+    make_nwbfile_from_metadata,
+    make_or_load_nwbfile,
+)
+from ._dataset_configuration import ConfigurableDataset, get_configurable_datasets
