+++ tests/test_minimal/test_tools/test_backend_and_dataset_configuration/test_get_configurable_datasets.py
@@ -8,7 +8,7 @@ from pynwb.base import DynamicTable
 from pynwb.testing.mock.base import mock_TimeSeries
 from pynwb.testing.mock.file import mock_NWBFile
 
-from neuroconv.tools.nwb_helpers import Dataset, get_io_datasets
+from neuroconv.tools.nwb_helpers import ConfigurableDataset, get_configurable_datasets
 
 
 def generate_1d_array() -> NWBFile:
@@ -21,7 +21,7 @@ def generate_2d_array() -> NWBFile:
     return array
 
 
-def generate_nwbfile_with_datasets() -> NWBFile:
+def generate_nwbfile_with_ConfigurableDatasets() -> NWBFile:
     nwbfile = mock_NWBFile()
     array = generate_2d_array()
     time_series = mock_TimeSeries(data=array)
@@ -31,9 +31,9 @@ def generate_nwbfile_with_datasets() -> NWBFile:
 
 @pytest.fixture(scope="session")
 def hdf5_nwbfile_path(tmpdir_factory):
-    nwbfile_path = tmpdir_factory.mktemp("data").join("test_hdf5_nwbfile_with_datasets.nwb")
+    nwbfile_path = tmpdir_factory.mktemp("data").join("test_hdf5_nwbfile_with_configurable_datasets.nwb.h5")
     if not Path(nwbfile_path).exists():
-        nwbfile = generate_nwbfile_with_datasets()
+        nwbfile = generate_nwbfile_with_ConfigurableDatasets()
         with NWBHDF5IO(path=str(nwbfile_path), mode="w") as io:
             io.write(nwbfile)
     return str(nwbfile_path)
@@ -41,9 +41,9 @@ def hdf5_nwbfile_path(tmpdir_factory):
 
 @pytest.fixture(scope="session")
 def zarr_nwbfile_path(tmpdir_factory):
-    nwbfile_path = tmpdir_factory.mktemp("data").join("test_zarr_nwbfile_with_datasets.nwb")
+    nwbfile_path = tmpdir_factory.mktemp("data").join("test_zarr_nwbfile_with_configurable_datasets.nwb.zarr")
     if not Path(nwbfile_path).exists():
-        nwbfile = generate_nwbfile_with_datasets()
+        nwbfile = generate_nwbfile_with_ConfigurableDatasets()
         with NWBZarrIO(path=str(nwbfile_path), mode="w") as io:
             io.write(nwbfile)
     return str(nwbfile_path)
@@ -56,12 +56,12 @@ def test_simple_time_series():
     time_series = mock_TimeSeries(name="TimeSeries", data=array)
     nwbfile.add_acquisition(time_series)
 
-    results = list(get_io_datasets(nwbfile=nwbfile))
+    results = list(get_configurable_datasets(nwbfile=nwbfile))
 
     assert len(results) == 1
 
     result = results[0]
-    assert isinstance(result, Dataset)
+    assert isinstance(result, ConfigurableDataset)
     assert result.object_name == "TimeSeries"
     assert result.field == "data"
     assert result.maxshape == array.shape
@@ -81,12 +81,12 @@ def test_simple_dynamic_table():
     dynamic_table.add_column(name="TestColumn", description="", data=array.squeeze())
     nwbfile.add_acquisition(dynamic_table)
 
-    results = list(get_io_datasets(nwbfile=nwbfile))
+    results = list(get_configurable_datasets(nwbfile=nwbfile))
 
     assert len(results) == 1
 
     result = results[0]
-    assert isinstance(result, Dataset)
+    assert isinstance(result, ConfigurableDataset)
     assert result.object_name == "TestColumn"
     assert result.field == "data"
     assert result.maxshape == (column_length,)
@@ -102,12 +102,12 @@ def test_simple_on_appended_hdf5_file(hdf5_nwbfile_path):
         new_time_series = mock_TimeSeries(name="NewTimeSeries", data=array)
         nwbfile.add_acquisition(new_time_series)
 
-        results = list(get_io_datasets(nwbfile=nwbfile))
+        results = list(get_configurable_datasets(nwbfile=nwbfile))
 
     assert len(results) == 1
 
     result = results[0]
-    assert isinstance(result, Dataset)
+    assert isinstance(result, ConfigurableDataset)
     assert result.object_name == "NewTimeSeries"
     assert result.field == "data"
     assert result.maxshape == array.shape
@@ -123,12 +123,12 @@ def test_simple_on_appended_zarr_file(zarr_nwbfile_path):
         new_time_series = mock_TimeSeries(name="NewTimeSeries", data=array)
         nwbfile.add_acquisition(new_time_series)
 
-        results = list(get_io_datasets(nwbfile=nwbfile))
+        results = list(get_configurable_datasets(nwbfile=nwbfile))
 
     assert len(results) == 1
 
     result = results[0]
-    assert isinstance(result, Dataset)
+    assert isinstance(result, ConfigurableDataset)
     assert result.object_name == "NewTimeSeries"
     assert result.field == "data"
     assert result.maxshape == array.shape
